+++ packages/layer-tests/src/e2e/services.rs
@@ -1,5 +1,7 @@
 use std::collections::BTreeMap;
 
+use crate::e2e::digests::DigestName;
+
 use super::{
     clients::Clients,
     config::Configs,
@@ -10,13 +12,22 @@ use alloy::sol_types::SolEvent;
 use utils::{
     context::AppContext,
     eigen_client::CoreAVSAddresses,
-    types::{ChainName, ComponentSource},
+    types::{
+        AllowedHostPermission, ChainName, Component, ComponentSource, Permissions, Service,
+        ServiceConfig, ServiceStatus, Submit, Trigger, Workflow,
+    },
+    ComponentID, ServiceID, WorkflowID,
 };
 use wavs_cli::{
     args::{CliSubmitKind, CliTriggerKind},
+    clients::example_eth_client::{
+        example_submit::SimpleSubmit, SimpleEthSubmitClient, SimpleEthTriggerClient,
+    },
     command::{
         deploy_eigen_core::{DeployEigenCore, DeployEigenCoreArgs},
+        deploy_eigen_service_manager::{DeployEigenServiceManager, DeployEigenServiceManagerArgs},
         deploy_service::{DeployService, DeployServiceArgs},
+        deploy_service_raw::{DeployServiceRaw, DeployServiceRawArgs},
     },
 };
 
@@ -24,7 +35,7 @@ use wavs_cli::{
 pub struct Services {
     #[allow(dead_code)]
     pub eth_eigen_core: BTreeMap<ChainName, CoreAVSAddresses>,
-    pub lookup: BTreeMap<AnyService, DeployService>,
+    pub lookup: BTreeMap<AnyService, Service>,
 }
 
 impl Services {
@@ -77,9 +88,15 @@ impl Services {
             let mut lookup = BTreeMap::default();
 
             // nonce errors here too, gotta go sequentially :/
-            for service in all_services {
-                let res = deploy_service(service, clients, digests, &chain_names).await;
-                lookup.insert(service, res);
+            for service_kind in all_services {
+                let service = match service_kind {
+                    AnyService::Eth(EthService::MultiWorkflow) => {
+                        deploy_service_raw(service_kind, clients, digests, &chain_names).await
+                    }
+                    _ => deploy_service_simple(service_kind, clients, digests, &chain_names).await,
+                };
+
+                lookup.insert(service_kind, service);
             }
 
             Self {
@@ -90,13 +107,14 @@ impl Services {
     }
 }
 
-async fn deploy_service(
+async fn deploy_service_simple(
     service: AnyService,
     clients: &Clients,
     digests: &Digests,
     chain_names: &ChainNames,
-) -> DeployService {
-    let digest = digests.lookup.get(&service.into()).unwrap().clone();
+) -> Service {
+    let digest_name = Vec::<DigestName>::from(service)[0];
+    let digest = digests.lookup.get(&digest_name).unwrap().clone();
 
     let trigger = match service {
         AnyService::Eth(_) => CliTriggerKind::EthContractEvent,
@@ -168,6 +186,140 @@ async fn deploy_service(
     .await
     .unwrap()
     .unwrap()
+    .service
+}
+
+async fn deploy_service_raw(
+    service_kind: AnyService,
+    clients: &Clients,
+    digests: &Digests,
+    chain_names: &ChainNames,
+) -> Service {
+    if !matches!(service_kind, AnyService::Eth(EthService::MultiWorkflow)) {
+        panic!("unexpected service kind: {:?}", service_kind);
+    }
+
+    let trigger1 = deploy_trigger_raw(clients, chain_names).await;
+    let trigger2 = deploy_trigger_raw(clients, chain_names).await;
+
+    let component_id1 = ComponentID::new("component1").unwrap();
+    let component_id2 = ComponentID::new("component2").unwrap();
+
+    let digest_names = Vec::<DigestName>::from(service_kind);
+
+    let component1 = Component {
+        wasm: digests.lookup.get(&digest_names[0]).unwrap().clone(),
+        permissions: Permissions {
+            allowed_http_hosts: AllowedHostPermission::All,
+            file_system: true,
+        },
+    };
+
+    let component2 = Component {
+        wasm: digests.lookup.get(&digest_names[1]).unwrap().clone(),
+        permissions: Permissions {
+            allowed_http_hosts: AllowedHostPermission::All,
+            file_system: true,
+        },
+    };
+
+    let submit1 = deploy_submit_raw(clients, chain_names).await;
+    let submit2 = deploy_submit_raw(clients, chain_names).await;
+
+    let workflow_id1 = WorkflowID::new("workflow1").unwrap();
+    let workflow_id2 = WorkflowID::new("workflow2").unwrap();
+
+    let workflow1 = Workflow {
+        trigger: trigger1,
+        component: component_id1,
+        submit: submit1,
+    };
+
+    let workflow2 = Workflow {
+        trigger: trigger2,
+        component: component_id2,
+        submit: submit2,
+    };
+
+    let components = BTreeMap::from([
+        (workflow1.component.clone(), component1),
+        (workflow2.component.clone(), component2),
+    ]);
+
+    let workflows = BTreeMap::from([(workflow_id1, workflow1), (workflow_id2, workflow2)]);
+
+    let service = Service {
+        id: ServiceID::new(uuid::Uuid::now_v7().as_simple().to_string()).unwrap(),
+        name: "".to_string(),
+        components,
+        workflows,
+        status: ServiceStatus::Active,
+        config: ServiceConfig::default(),
+        testable: true,
+    };
+
+    DeployServiceRaw::run(
+        &clients.cli_ctx,
+        DeployServiceRawArgs {
+            service: service.clone(),
+        },
+    )
+    .await
+    .unwrap();
+
+    service
+}
+
+async fn deploy_trigger_raw(clients: &Clients, chain_names: &ChainNames) -> Trigger {
+    let chain_name = chain_names.eth[0].clone();
+    let eigen_client = clients.cli_ctx.get_eth_client(&chain_name).unwrap().clone();
+    let event_hash =
+        *wavs_cli::clients::example_eth_client::example_trigger::NewTrigger::SIGNATURE_HASH;
+
+    let address = SimpleEthTriggerClient::deploy(eigen_client.eth.provider.clone())
+        .await
+        .unwrap();
+
+    Trigger::EthContractEvent {
+        chain_name,
+        address,
+        event_hash,
+    }
+}
+
+async fn deploy_submit_raw(clients: &Clients, chain_names: &ChainNames) -> Submit {
+    let chain_name = chain_names.eth[0].clone();
+    let eigen_client = clients.cli_ctx.get_eth_client(&chain_name).unwrap().clone();
+
+    let simple_submit = SimpleSubmit::deploy(eigen_client.eth.provider.clone())
+        .await
+        .unwrap();
+
+    let res = DeployEigenServiceManager::run(
+        &clients.cli_ctx,
+        DeployEigenServiceManagerArgs {
+            chain: chain_name.clone(),
+            service_handler: *simple_submit.address(),
+            register_operator: true,
+        },
+    )
+    .await
+    .unwrap();
+
+    let DeployEigenServiceManager { address, .. } = res;
+    let simple_submit_client =
+        SimpleEthSubmitClient::new(eigen_client.eth.clone(), *simple_submit.address());
+
+    simple_submit_client
+        .set_service_manager_address(address)
+        .await
+        .unwrap();
+
+    Submit::EigenContract {
+        chain_name,
+        service_manager: address,
+        max_gas: None,
+    }
 }
 
 #[derive(Default)]