+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning].
 
 - CLI supports "none" for submit kind
 - WAVS namespaces storage only by ServiceID, not ServiceID + WorkflowID
+- WAVS and CLI support raw service deployment
+- Overall support for multi-workflow services (including e2e tests)
 
 ## [v0.3.0-alpha3]